Separated ted.c in ted.c, ted.h and message_and_prompt.c
