+++ Makefile
@@ -1,10 +1,11 @@
 all: ted
 
-ted: ted.c
-	$(CC) -o ted ted.c -lncursesw -Wall -W -pedantic -std=c99
+ted: src/*
+	$(CC) -o ted src/* -lncursesw -Wall -W -pedantic -std=c99
 
-dev: ted.c
-	$(CC) -o ted ted.c -lncursesw -Wall -W -pedantic -std=c99 -fsanitize=address
+dev: src/*
+	$(CC) -o ted src/* -lncursesw -Wall -W -pedantic -std=c99 -fsanitize=address
 
 clean:
 	rm ted
+
