--- tips.txt
@@ -1,3 +1,0 @@
-C0---DF
-E0---EF
-F0---F7
